pinokio_AI-main/requirements.txt
protobuf
fastapi
uvicorn
google-genai
pydantic
python-dotenv
